style: Move uart code to a separate module
